Weight normalization for RNN Cells.
The current RNN implementation executes a user defined
function (the call() method of subclasses of RNNCells) inside a
tf.while() loop. Weight normalisation requires a one-time
 normalization of the transition matrices prior to
entering the while loop. The following 2 edits have been made in
tensorflow/python/ops to enable this functionality:

 - RNNCell now has a prepare() method. It does nothing as implemented
in the base class
 - A call to cell.prepare() has been added just before entering
 _dynamic_rnn_loop()

Subclasses of RNNCell may implement normalization in the cell's
prepare() method. One implementation with BasicLSTMCell
and associated tests have been added to contrib. Note that any
wrappers to be used with a weight-normalized cell need to be
appropriately subclassed, as illustrated with the
PrepareableMultiRNNCell example in contrib.